+++ packages/moneystream-extension/src/popup/Index.tsx
@@ -9,9 +9,10 @@ import { AccountBar } from './components/AccountBar'
 import { WebMonetizedBar } from './components/WebMonetizedBar'
 import { Status } from './components/Status'
 import { PayToUrlPage } from './components/PayToUrl'
+import { HistoryPage } from './components/History'
 import { PopupProps } from './types'
 import { MemoryRouter, Switch, Route } from 'react-router-dom'
-import NavBar from './components/NavBar'
+import { NavBar } from './components/NavBar'
 
 const MoneystreamContainer = styled(Container)(({ theme }) => ({
   paddingRight: `${theme.spacing(4)}px`,
@@ -57,16 +58,17 @@ export function Index(props: PopupProps) {
       <MoneystreamContainer>
         {/* <Status context={context} /> */}
         <MemoryRouter>
-          <NavBar />
           <Switch>
             <Route path='/status' component={() => <Status context={context}/>} />
+            <Route path='/history' component={() => <HistoryPage context={context}/>} />
             <Route path='/paytourl' component={() => <PayToUrlPage context={context}/>} />
             {/* <Route path='/settings' component={Settings} /> */}
             <Route path='/' component={() => <Status context={context}/>} />
           </Switch>
+          <WebMonetizedBar context={context} />
+          <NavBar context={context} />
         </MemoryRouter>
       </MoneystreamContainer>
-      <WebMonetizedBar context={context} />
     </OuterDiv>
   )
 }